week6ListDictionaries
Write a program and continuously ask the user to enter different names, until the user presses Enter (without supplying a name). Depending on the number of names provided, display a message based on the above pattern. You should use a List when solving this one!

Write a program that asks the user to enter a sentence. Display each unique letter in the sentence, and how many times each one appeared. You should use a Dictionary when solving this one!